Fix: 아이디 변수 타입 불일치 수정

+++ backend/dagaga/domain/src/main/java/com/dagaga/domain/user/dto/UserRegisterDto.java
@@ -27,6 +27,6 @@ public class UserRegisterDto {
     private String nickname;
     private String viewLangCode;
     private String nativeLangCode;
-    private Long locationId;
+    private Integer locationId;
     private LocalDate arrivalDate;
 }
